requirements.txt
# AI Development Platform Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11

# Database


# AI and ML
openai==1.3.7
gpt-engineer==0.3.1

# Blockchain
solana==0.30.2
base58==2.1.1

# Data Processing
pydantic==2.5.0
pyyaml==6.0.1

# HTTP and API
requests==2.31.0
httpx==0.23

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.3.0
flake8==6.1.0

# Optional: For enhanced features
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0